NemoNemoLogic-master/requirements.txt
pyqt5==5.15.10
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
 